게시글 첨부파일 back-end

+++ finalproject/src/main/java/com/kh/finalproject/dto/AttachmentDto.java
@@ -11,9 +11,9 @@ import lombok.NoArgsConstructor;
 
 @Data @NoArgsConstructor @AllArgsConstructor @Builder
 public class AttachmentDto {
-	private Integer attachmentNo; //시퀀스로 생성(첨부파일 번호)
+	private int attachmentNo; //시퀀스로 생성(첨부파일 번호)
 	private String attachmentName;
-	private Integer attachmentSize;
+	private Long attachmentSize;
 	private String attachmentType;
 	@JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
 	private LocalDateTime attachmentTime;
